Resubmitting waterregulation
I'm not getting the same pylint errors on my files. I think there's a version control issue that I don't know how to fix. It looks like the files that have the errors are the ones in the pump and sensor folders, but on my last commit I moved sensor and pump to the main folder with a single init file. I haven't been able to delete the sensor and pump folders, however. Also, I'm not sure why, but the .pylintrc file won't upload to git. I'll send via email.